+++ src/main/java/com/barberapp/services/barber/ServiceBarbermpl.java
@@ -36,7 +36,7 @@ public class ServiceBarbermpl implements ServiceBarber{
 	
 
 	@Transactional(readOnly = true)
-	public Optional<Barber> findById(Integer id) {
+	public Optional<Barber> findById(Long id) {
 
 		return barber.findById(id);
 	}
@@ -49,7 +49,7 @@ public class ServiceBarbermpl implements ServiceBarber{
 	}
 
 	@Override
-	public void deletById(Integer id) {
+	public void deletById(Long id) {
 		barber.deleteById(id);
 		
 	}
